rename: features, entities 공통 사용 컴포넌트 위치 변경
