+++ src/entities/admin/common/RichEditor.ts
